% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{allMarkers}
\alias{allMarkers}
\title{Generate markers of Seurat identity classes and order them decreasingly}
\usage{
allMarkers(seuratObj, objectName = NULL, group.by = "seurat_clusters", ...)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{objectName}{Character to be used in the file name when saving the markers.
Default is NULL (the markers will not be saved)}

\item{group.by}{Regroup cells into a different identity class prior to 
performing differential expression (see example); \code{"ident"} to use Idents}

\item{...}{Arguments passed to other methods and to specific DE methods}
}
\value{
A data frame of markers ordered decreasingly by pct.ratio (pct.1 / pct.2)
}
\description{
This function is a wrapper around FindMarkers from Seurat that relaxes the
criteria for marker selection as default option, calculates pct.ratio (pct. 1
 / pct.2), and used it to order decreasingly the data frame.
}
