% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markerListNames}
\alias{markerListNames}
\title{Generate list of marker names from list of marker data frames}
\usage{
markerListNames(
  markerList,
  removeRep = FALSE,
  percMarkers = 2,
  nMarkersThr = 5
)
}
\arguments{
\item{markerList}{A data frame generated using FindMarkers or orderedFindMarkers
The latter is required if pct.ratio is listed among the criteria}

\item{removeRep}{If true, remove the markers appearing in more than one data
frame. Default is FALSE}

\item{percMarkers}{Percentage of top markers retained. If NULL, all markers
will be retained}

\item{nMarkersThr}{The lowest acceptable number of markers in a data frame;
if a data frame has fewer marker than this number, its markers will be
excluded from the output}
}
\value{
A data frame of markers ordered decreasingly by pct.ratio (pct.1 / pct.2)
}
\description{
This function takes a list of marker data frames generated using FindMarkers
or orderedFindMarkers, ranks and filters the markers in each data frame,
and returns a list of marker names.
}
