% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{orderedFindMarkers}
\alias{orderedFindMarkers}
\title{Generate markers of Seurat identity classes and order them decreasingly}
\usage{
orderedFindMarkers(
  seuratObj,
  min.pct = 0,
  logfc.threshold = 0,
  min.diff.pct = 0,
  only.pos = T,
  densify = T,
  min.cells.group = 1,
  ...
)
}
\arguments{
\item{seuratObj}{A Seurat object}

\item{min.pct}{only test genes that are detected in a minimum fraction of
min.pct cells in either of the two populations. Meant to speed up the function
by not testing genes that are very infrequently expressed. Default is 0.01}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least
X-fold difference (log-scale) between the two groups of cells. Default is 0.1
Increasing logfc.threshold speeds up the function, but can miss weaker signals.
If the \code{slot} parameter is "scale.data" no filtering is performed.}

\item{min.diff.pct}{only test genes that show a minimum difference in the
fraction of detection between the two groups. Set to -Inf by default}

\item{only.pos}{Only return positive markers (FALSE by default)}

\item{densify}{Convert the sparse matrix to a dense form before running the 
DE test. This can provide speedups but might require higher memory; default is FALSE}

\item{min.cells.group}{Minimum number of cells in one of the groups}

\item{...}{Arguments passed to other methods and to specific DE methods}
}
\value{
A data frame of markers ordered decreasingly by pct.ratio (pct.1 / pct.2)
}
\description{
This function is a wrapper around FindMarkers from Seurat that relaxes the
criteria for marker selection as default option, calculates pct.ratio (pct. 1
 / pct.2), and used it to order decreasingly the data frame.
}
