% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{rankMarkers}
\alias{rankMarkers}
\title{Rank markers of Seurat identity classes}
\usage{
rankMarkers(
  markers,
  criteria = c("avg_log2FC", "pct.1", "p_val_adj", "pct.ratio"),
  signs = c(-1, -1, 1, -1),
  percMarkers = 2,
  rankFun = MatrixGenerics::rowMaxs
)
}
\arguments{
\item{markers}{A data frame generated using FindMarkers or orderedFindMarkers
The latter is required if pct.ratio is listed among the criteria}

\item{criteria}{Columns of the input data frame that will be involved in the
ranking. Must be character objects}

\item{signs}{A vector of the same lengths as criteria, in which -1 and 1 are
the only allowed values; -1 indicates that the corresponding criterion will
be used to rank the markers decreasingly, while 1 indicates that it will be
used to rank the markers increasingly
Default is NULL (the markers will not be saved)}

\item{percMarkers}{Percentage of top markers retained. If NULL, all markers
will be retained}

\item{rankFun}{The function used to rank markers}
}
\value{
A data frame of markers ordered decreasingly by pct.ratio (pct.1 / pct.2)
}
\description{
This function ranks markers generated using FindMarkers or orderedFindMarkers
using input criteria, and filters top markers using a percentage argument.
}
