% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{ratioOrder}
\alias{ratioOrder}
\title{Order markers decreasingly by ratio}
\usage{
ratioOrder(markers)
}
\arguments{
\item{markers}{A data frame generated using FindMarkers}
}
\value{
A data frame of markers ordered decreasingly by pct.ratio (pct.1 / pct.2)
}
\description{
This function calculates pct.ratio (pct.1 / pct.2) for a data frame generated
using FindMarkers, then used it to order the data frame decreasingly.
}
