% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{removeRepeatedMarkers}
\alias{removeRepeatedMarkers}
\title{Remove non-exclusive markers}
\usage{
removeRepeatedMarkers(markerList)
}
\arguments{
\item{markerList}{A data frame generated using FindMarkers or orderedFindMarkers
The latter is required if pct.ratio is listed among the criteria}
}
\value{
A list of marker data frames containing only exclusive markers
}
\description{
This function removes all markers found in more than one data frame from an
input list
}
